Remove unnecessary definitions for Elucidation
